Rename `flow` package to `transformation` (#137)
