+++ app/transformation/AbstractTripleRematch.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package flow;
+package transformation;
 
 import java.io.File;
 import java.io.FileOutputStream;
