+++ app/transformation/Enrich.java
@@ -1,4 +1,4 @@
-package flow;
+package transformation;
 
 import java.io.File;
 import java.io.IOException;
