+++ app/transformation/package-info.java
@@ -3,4 +3,4 @@
  * 
  * @author Fabian Steeg (fsteeg)
  */
-package flow;
+package transformation;
